+++ src/components/Panes/AboutLanguage/AboutLanguage.tsx
@@ -12,6 +12,7 @@ import AboutLanguageIntro from './AboutLanguageIntro';
 import AboutLanguageLetters from './AboutLanguageLetters';
 import AboutLanguageNeighbours from './AboutLanguageNeighbours';
 import AboutLanguagePlayDiffle from './AboutLanguagePlayDiffle';
+import AboutLanguageWordle from './AboutLanguageWordle';
 import AboutLenguageLengths from './AboutLenguageLengths';
 
 import './AboutLanguage.scss';
@@ -23,6 +24,15 @@ const getDicitonaryData = async (lang: string | undefined): Promise<DictionaryIn
 
   const response = await fetch(`./dictionary/${lang}/info.json`);
   const rawData = await response.json();
+
+  const responseWordleBest = await fetch(`./dictionary/${lang}/info-wordle-best.json`);
+  const rawDataWordleBest = await responseWordleBest.json();
+
+  rawData.spellchecker.wordle.bestMax = rawDataWordleBest.best.max;
+  rawData.spellchecker.wordle.bestMaxGreen = rawDataWordleBest.best.maxGreen;
+  rawData.spellchecker.wordle.bestMaxOrange = rawDataWordleBest.best.maxOrange;
+  rawData.spellchecker.wordle.bestGreen1_5 = rawDataWordleBest.best.green1_5;
+  rawData.spellchecker.wordle.bestGreen2_0 = rawDataWordleBest.best.green2_0;
 
   return rawData;
 };
@@ -50,6 +60,7 @@ const AboutLanguage = () => {
               <AboutLanguageLetters data={data} groupBy={DictionaryInfoLetters.InWords} />
               <AboutLanguageNeighbours data={data} />
               <AboutLenguageLengths data={data} />
+              <AboutLanguageWordle data={data} />
               <AboutLanguagePlayDiffle />
           </>
           )}