+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 ---
 
+#### Version 3.9
+- **New**: Section about language added with the best starting word for Wordle
+- **New**: The length of the word has been increased in the About Language section
+
 #### Version 3.8.2 (24.02.2024)
 - **Fix**: Shared content is displayed again